# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CoinswapApi V1 REST API Documentation",
        "description": "This is the documentation for the V1 of the rest API",
        "contact": {
            "email": "rafiumoshoodolakunle@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/balance/all": {
            "get": {
                "tags": [
                    "Balance"
                ],
                "summary": "get user balance for all supported coins",
                "operationId": "get user balance for all supported coins",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/balance/{coin}": {
            "get": {
                "tags": [
                    "Balance"
                ],
                "summary": "get user balance by a specific coin short name supported coins",
                "operationId": "get user balance by a specific coin short name supported coins",
                "parameters": [
                    {
                        "name": "coin",
                        "in": "path",
                        "description": "e.g btc, eth, ltc",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}